Move HeadersMap to types folder
