+++ src/Api/Api.ts
@@ -1,4 +1,6 @@
-import { Method, HeadersMap, ApiResponse } from './types';
+import { HeadersMap } from '../types';
+
+import { Method, ApiResponse } from './types';
 import createRequest from './createRequest';
 
 interface Params {
