+++ src/Api/ApiError.ts
@@ -1,4 +1,6 @@
-import { ApiErrorPayload, HeadersMap, ApiResponse } from './types';
+import { HeadersMap } from '../types';
+
+import { ApiErrorPayload, ApiResponse } from './types';
 
 export default class ApiError<P = ApiErrorPayload> extends Error implements ApiResponse<P> {
     payload: P;
