+++ src/types/HeadersMap.ts
@@ -1,0 +1,3 @@
+export default interface HeadersMap {
+    [name: string]: string;
+}
