+++ src/types/index.ts
@@ -1,6 +1,7 @@
 export { default as Contact } from './Contact';
 export { default as Coverage } from './Coverage';
 export { default as Entity } from './Entity';
+export { default as HeadersMap } from './HeadersMap';
 export { default as Pagination } from './Pagination';
 export { default as RoomRef } from './RoomRef';
 export { default as Story } from './Story';
